Consolidate demand UGens
