--- supriya/ugens/Dshuf.py
@@ -1,29 +1,0 @@
-import collections
-
-from supriya import CalculationRate
-from supriya.ugens.DUGen import DUGen
-
-
-class Dshuf(DUGen):
-    """
-    A demand-rate random sequence generator.
-
-    ::
-
-        >>> sequence = (1, 2, 3)
-        >>> dshuf = supriya.ugens.Dshuf.new(
-        ...     repeats=1,
-        ...     sequence=sequence,
-        ...     )
-        >>> dshuf
-        Dshuf()
-
-    """
-
-    ### CLASS VARIABLES ###
-
-    _ordered_input_names = collections.OrderedDict([("repeats", 1), ("sequence", None)])
-
-    _unexpanded_input_names = ("sequence",)
-
-    _valid_calculation_rates = (CalculationRate.DEMAND,)
